requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
typer==0.12.5
click==8.1.7
rich==13.9.1
pydantic==2.9.2
python-dateutil==2.9.0.post0
fastapi==0.115.0
uvicorn==0.30.6
Jinja2==3.1.4
python-dotenv==1.0.1
